per review, moved to xclim.indices.solar, merged with xclim.indices.helpers
