+++ src/xclim/indices/converters.py
@@ -27,13 +27,15 @@ from xclim.core.units import (
 from xclim.indices.helpers import (
     _gather_lat,
     _gather_lon,
+    wind_speed_height_conversion,
+)
+from xclim.indices.solar import (
     cosine_of_solar_zenith_angle,
     day_lengths,
     distance_from_sun,
     extraterrestrial_solar_radiation,
     solar_declination,
     time_correction_for_solar_angle,
-    wind_speed_height_conversion,
 )
 
 __all__ = [
