+++ tests/test_atmos.py
@@ -2,7 +2,7 @@ import numpy as np
 import pytest
 
 from xclim import atmos, set_options
-from xclim.indices.helpers import make_hourly_temperature
+from xclim.indices.solar import make_hourly_temperature
 
 K2C = 273.16
